+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 1.0.5
 =====
+
 *   (improvement) Extract `filterOptionalFields` to `FilterFieldsHelper`.
 *   (bug) Add missing filter for optional fields in `Slice`.
+*   (bug) Set `repeat` or `non-repeat` in `Slice` to `null`, if empty.
 
 
 1.0.4