+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 *   (improvement) Extract `filterOptionalFields` to `FilterFieldsHelper`.
 *   (bug) Add missing filter for optional fields in `Slice`.
 *   (bug) Set `repeat` or `non-repeat` in `Slice` to `null`, if empty.
+*   (bug) Fix wrong value of `[config][labels]` in `SliceZone`.
 
 
 1.0.4